move models to models dir
