+++ app/models/forms.py
@@ -1,10 +1,13 @@
 """
 Forms for registering doctors, patients, and prescriptions.
 """
-
 from flask_wtf import FlaskForm
 from flask_wtf.file import FileRequired
-from wtforms import FileField, PasswordField, StringField, SubmitField, IntegerField
+from wtforms import FileField
+from wtforms import IntegerField
+from wtforms import PasswordField
+from wtforms import StringField
+from wtforms import SubmitField
 from wtforms.validators import DataRequired
