+++ app/models/persons.py
@@ -1,11 +1,10 @@
 """
 Classes to model users (doctors) and patients.
 """
-
 from collections import defaultdict
 from datetime import datetime
-import numpy as np
 
+import numpy as np
 from flask_login import UserMixin
 from werkzeug.security import check_password_hash
 from werkzeug.security import generate_password_hash
@@ -106,14 +105,14 @@ class Patient(db.Model):
         self, on_time_threshold=0.9, required_intakes_threshold=0.9
     ):
         """
-        Return fraction of prescriptions for this Patient that are deemed 
+        Return fraction of prescriptions for this Patient that are deemed
         adherent.
-        Adherence requires Intakes to be recorded on time and on track 
+        Adherence requires Intakes to be recorded on time and on track
         (i.e. medications aren't missed).
 
-        on_time_threshold: float - fraction on time Intakes needed for 
+        on_time_threshold: float - fraction on time Intakes needed for
         this prescription to be deemed adherent
-        required_intakes_threshold: float - fraction Intakes actually 
+        required_intakes_threshold: float - fraction Intakes actually
         recorded, out of all prescribed Intakes since start date.
         """
         if len(self.prescriptions) == 0:
@@ -139,12 +138,12 @@ class Patient(db.Model):
         self, on_time_threshold=0.9, required_intakes_threshold=0.9, date=datetime.now()
     ):
         """
-        Whether or not a patient is deemed adherent based on their 
+        Whether or not a patient is deemed adherent based on their
         prescription adherence.
 
-        on_time_threshold: float - fraction on time Intakes needed for 
+        on_time_threshold: float - fraction on time Intakes needed for
         this prescription to be deemed adherent
-        required_intakes_threshold: float - fraction Intakes actually 
+        required_intakes_threshold: float - fraction Intakes actually
         recorded, out of all prescribed Intakes since start date.
         """
         if np.all([p.start_date > datetime.now() for p in self.prescriptions]):
